# docs/openapi/openapi-v1.yaml
openapi: 3.0.3
info:
  title: TechDengue — API v1
  version: 1.0.0
  description: |
    API **Edital-Core++** para indicadores EPI, operação de campo, evidências e relatórios.
    Segurança via **OAuth2/OIDC** (Bearer JWT). Todas as rotas v1 retornam `application/json`
    por padrão, exceto downloads (PDF/CSV).
servers:
  - url: https://api.techdengue.local/v1
    description: Ambiente de homologação (exemplo)
  - url: http://localhost:8080/v1
    description: Local dev
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    MunicipioIBGE:
      in: query
      name: municipio_cod_ibge
      schema: { type: string, pattern: "^[0-9]{7}$" }
      required: true
      description: Código IBGE do município (7 dígitos).
    Competencia:
      in: query
      name: competencia
      schema: { type: string, format: date }
      required: true
      description: Data da competência (usar último dia do mês).
...
  schemas:
...
    RelatorioLinks:
      type: object
      properties:
        pdf_url: { type: string }
        csv_url: { type: string }
        hash_sha256: { type: string }
...
paths:
...
  /relatorios/epi01:
    get:
      tags: [Relatorios]
      summary: Gera/retorna links para EPI01 (PDF/CSV)
      parameters:
        - $ref: "#/components/parameters/MunicipioIBGE"
        - $ref: "#/components/parameters/Competencia"
        - in: query
          name: formato
          schema: { type: string, enum: [pdf, csv] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RelatorioLinks" }
...